Extracting AlchemyHttpImpl to Outer Scope
